Patch: Corrected backwards iteration logic on event handler chain.

+++ src/MWindow.cpp
@@ -100,7 +100,7 @@ namespace MW {
     void MWindow::executeHandlerChain(const MEvent& ev) {
         std::lock_guard<std::mutex> lock(handler_lock);
         if(!handlers.size()) return;
-        for(size_t i{handlers.size()-1};i >= 0;--i)
+        for(int i{static_cast<int>(handlers.size()-1)};i >= 0;--i)
         {
             if(handlers[i].handler(ev) == MEventResult::Consumed)
                 break;
